# internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Bekci API
  version: "1"
  description: |
    Bekci'nin (kendi sunucunuzda çalışan uptime ve sunucu izleme) REST API'si.
    Arayüzün kullandığı API ile aynıdır; betikler ve Terraform/Ansible gibi
    araçlar **API anahtarı** ile kullanır.

    **Kimlik doğrulama:** `Authorization: Bearer upk_…` (Ayarlar → Hesabım →
    API anahtarları). Tarayıcı oturumu (çerez) ile çağrılan yazma istekleri
    `X-Uptime: 1` başlığı ister (CSRF); API anahtarıyla gerek yoktur.
    Hesap güvenliği uçları (`/api/auth/*`, kullanıcı yönetimi) API anahtarıyla
    kullanılamaz.

    **Roller:** izleyici (viewer) okur; editör (editor) monitör, kanal, durum
    sayfası ve bakım yönetir; yönetici (admin) her şeyi. Yanıt dili
    `X-Uptime-Lang: tr|en` ya da `Accept-Language` ile seçilir.

    **Zamanlar** Unix saniyedir (`started_at`, `created_at` …); `0` "yok" demektir.
    Hatalar `{"error": "…", "code": "…"}` gövdesiyle döner (code isteğe bağlı).
  license:
    name: AGPL-3.0
    url: https://github.com/kadirsungurlu/bekci/blob/main/LICENSE
servers:
  - url: /
security:
  - apiKey: []
  - session: []
...
components:
  requestBodies:
    Credentials:
      content:
        application/json:
          schema:
            type: object
            properties:
              username: { type: string }
              password: { type: string }
            required: [username, password]
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: API anahtarı (`upk_…`); rolü anahtarın rolü ile sahibinin rolünün küçüğüdür.
    session:
      type: apiKey
      in: cookie
      name: uptime_oturum
      description: "Tarayıcı oturumu; yazma isteklerinde `X-Uptime: 1` başlığı gerekir."
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    Error:
      description: Hata
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    OK:
      description: Başarılı
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OK" }
  schemas:
...
    Lang:
      type: string
      enum: [tr, en]
    Theme:
      type: string
      enum: ["", light, dark]
    User:
      type: object
      description: Oturumdaki kullanıcının görünümü (`/api/auth/*`).
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        display_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        must_change_password: { type: boolean }
        all_monitors: { type: boolean }
        two_factor_enabled: { type: boolean }
        servers: { type: boolean, description: Sunucular ekranı açık mı }
        lang: { type: string }
        email: { type: string }
        oidc: { type: boolean, description: Bir SSO sağlayıcısına bağlı }
        theme: { $ref: "#/components/schemas/Theme" }
...
    Bucket:
      type: object
      description: Saatlik/günlük özet kovası.
      properties:
        t: { type: integer, format: int64 }
        up: { type: integer }
        down: { type: integer }
        ping: { type: number, description: "Ortalama yanıt (ms); -1 yok" }
        ping_min: { type: number }
        ping_max: { type: number }
    Monitor:
      type: object
      description: Monitör ayarları ve son durumu (liste görünümü ek alanlar taşır).
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        type:
          type: string
          enum: [http, tcp, ping, dns, tlscert, smtp, websocket, grpc, mqtt, snmp, mysql, postgres, mssql, redis, mongodb, docker, push, group]
        description: { type: string }
        active: { type: boolean }
        interval: { type: integer, description: Kontrol aralığı (sn) }
        retry_interval: { type: integer }
        max_retries: { type: integer }
        timeout: { type: integer }
        resend_every: { type: integer, description: "Hatırlatma: her N başarısız kontrolde (0 kapalı)" }
        upside_down: { type: boolean }
        config: { type: object, additionalProperties: true, description: Tipe özgü ayarlar (gizli alanlar maskeli) }
        push_token: { type: string, description: Yalnızca push tipinde (editör+) }
        slow_ms: { type: integer }
        slow_checks: { type: integer }
        slow: { type: boolean }
        status: { type: integer, description: "0 çalışmıyor, 1 çalışıyor, 2 bekliyor, 3 bakımda" }
        last_check_at: { type: integer, format: int64 }
        last_change_at: { type: integer, format: int64 }
        last_ping_ms: { type: integer, format: int64 }
        last_message: { type: string }
        cert_expires_at: { type: integer, format: int64 }
        cert_issuer: { type: string }
        domain_expiry: { type: boolean }
        domain_name: { type: string }
        domain_expires_at: { type: integer, format: int64 }
        domain_registrar: { type: string }
        domain_status: { type: string }
        created_at: { type: integer, format: int64 }
        updated_at: { type: integer, format: int64 }
        target: { type: string, description: Okunur hedef (adres/host) }
        notification_ids: { type: array, items: { type: integer, format: int64 } }
        tag_notification_ids: { type: array, items: { type: integer, format: int64 }, description: Etiket kuralıyla bağlı kanallar }
        tags: { type: array, items: { $ref: "#/components/schemas/MonitorTag" } }
        uptime_24h: { type: [number, "null"] }
        uptime_7d: { type: [number, "null"] }
        uptime_30d: { type: [number, "null"] }
        bars: { type: array, items: { $ref: "#/components/schemas/Bucket" } }
        pings: { type: array, items: { type: integer } }
        in_maintenance: { type: boolean }
        locations: { type: object, additionalProperties: true, description: Kontrol konumları (include_local, probe_ids, down_when, notify_partial) }
        open_incident_id: { type: [integer, "null"] }
        open_partial_incident_id: { type: [integer, "null"] }
        open_degraded_incident_id: { type: [integer, "null"] }
...
paths:
...
  /api/auth/setup:
    post:
      tags: [auth]
      summary: İlk yönetici hesabı (yalnızca hiç kullanıcı yokken)
      security: []
      requestBody: { $ref: "#/components/requestBodies/Credentials" }
      responses:
        "200": { description: Oturum açıldı, content: { application/json: { schema: { type: object, properties: { user: { $ref: "#/components/schemas/User" } } } } } }
        "409": { $ref: "#/components/responses/Error" }
  /api/auth/login:
    post:
      tags: [auth]
      summary: Şifreyle giriş
      description: "2FA açıksa `{\"two_factor_required\": true, \"challenge\": \"…\"}` döner; `/api/auth/login/2fa` ile tamamlanır. IP başına 15 dk'da 5, hesap başına 50 hatalı deneme sınırı."
      security: []
      requestBody: { $ref: "#/components/requestBodies/Credentials" }
      responses:
        "200": { description: Oturum ya da 2FA adımı }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
  /api/auth/password:
    post:
      tags: [auth]
      summary: Kendi şifresini değiştir (diğer oturumlar kapanır)
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { current: { type: string }, new: { type: string, minLength: 8 } }, required: [current, new] }
      responses: { "200": { $ref: "#/components/responses/OK" }, "400": { $ref: "#/components/responses/Error" } }
  /api/auth/preferences:
    put:
      tags: [auth]
      summary: Arayüz dili ve tema tercihi
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { lang: { type: string, enum: ["", tr, en] }, theme: { $ref: "#/components/schemas/Theme" } } }
      responses: { "200": { description: Güncel kullanıcı, content: { application/json: { schema: { type: object, properties: { user: { $ref: "#/components/schemas/User" } } } } } } }
...
  /api/monitors:
    get:
      tags: [monitors]
      summary: Monitörler (kısıtlı kullanıcıda yalnızca izinliler)
      responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Monitor" } } } } } }
    post:
      tags: [monitors]
      summary: Monitör ekle (editör+)
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/MonitorInput" } } } }
      responses: { "201": { description: Monitör, content: { application/json: { schema: { $ref: "#/components/schemas/Monitor" } } } }, "400": { $ref: "#/components/responses/Error" } }
...
  /api/monitors/{id}:
    parameters: [{ $ref: "#/components/parameters/id" }]
    get:
      tags: [monitors]
      summary: Monitör ayrıntısı
      responses: { "200": { description: "{monitor, uptime{24h,7d,30d,90d}, avg_ping_24h, open_incident_since, open_incident_id}" }, "404": { $ref: "#/components/responses/Error" } }
    put:
      tags: [monitors]
      summary: Monitörü güncelle (editör+)
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/MonitorInput" } } } }
      responses: { "200": { description: Monitör }, "400": { $ref: "#/components/responses/Error" } }
    delete:
      tags: [monitors]
      summary: Monitörü sil (editör+)
      responses: { "200": { $ref: "#/components/responses/OK" } }
  /api/monitors/{id}/pause:
    post: { tags: [monitors], summary: Durdur (editör+), parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { $ref: "#/components/responses/OK" } } }
  /api/monitors/{id}/resume:
    post: { tags: [monitors], summary: Başlat (editör+), parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { $ref: "#/components/responses/OK" } } }
...
  /api/monitors/{id}/series:
    get:
      tags: [monitors]
      summary: Grafik serisi
      parameters: [{ $ref: "#/components/parameters/id" }, { name: range, in: query, schema: { type: string, enum: [24h, 7d, 30d, 90d] } }]
      responses: { "200": { description: Kovalar, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Bucket" } } } } } }
  /api/monitors/{id}/beats:
    get:
      tags: [monitors]
      summary: Son kontroller (yeniden eskiye)
      parameters: [{ $ref: "#/components/parameters/id" }, { name: limit, in: query, schema: { type: integer, maximum: 200, default: 50 } }]
      responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Beat" } } } } } }
  /api/monitors/{id}/incidents:
    get: { tags: [monitors], summary: Monitörün olayları, parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Incident" } } } } } } }
...
  /api/monitors/{id}/notifications:
    put:
      tags: [monitors]
      summary: Bildirim kanallarını ayarla (editör+)
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { notification_ids: { type: array, items: { type: integer, format: int64 } } } } } } }
      responses: { "200": { description: "{notification_ids}" } }
...
  /api/incidents:
    get:
      tags: [incidents]
      summary: Olaylar (süzgeçler ve sayfalama; format=csv ile CSV)
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: ["", monitor, partial, server, degraded, manual] } }
        - { name: open, in: query, schema: { type: string, enum: ["", "1"] }, description: Yalnızca sürenler }
        - { name: monitor_id, in: query, schema: { type: integer } }
        - { name: server_id, in: query, schema: { type: integer } }
        - { name: from, in: query, schema: { type: integer }, description: Başlangıç ≥ (unix) }
        - { name: to, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
        - { name: before, in: query, schema: { type: integer }, description: "Sayfalama: bu kimlikten küçükler" }
        - { name: limit, in: query, schema: { type: integer, maximum: 500 } }
        - { name: format, in: query, schema: { type: string, enum: [csv] } }
      responses:
        "200":
          description: Liste ya da CSV
          content:
            application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Incident" } } }
            text/csv: { schema: { type: string } }
    post:
      tags: [incidents]
      summary: Elle olay aç (durum sayfası iletişimi; editör+)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                page_id: { type: integer, format: int64 }
                title: { type: string }
                severity: { type: string, enum: [minor, major, critical] }
                monitor_ids: { type: array, items: { type: integer, format: int64 } }
                started_at: { type: integer, format: int64 }
                state: { type: string }
                body: { type: string }
              required: [page_id, title]
      responses: { "201": { description: Olay, content: { application/json: { schema: { $ref: "#/components/schemas/Incident" } } } } }
...
  /api/notifications:
    get: { tags: [notifications], summary: Bildirim kanalları (editör+; gizli alanlar maskeli), responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Notification" } } } } } } }
    post:
      tags: [notifications]
      summary: Kanal ekle (editör+)
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/NotificationInput" } } } }
      responses: { "201": { description: Kanal }, "400": { $ref: "#/components/responses/Error" } }
  /api/notifications/{id}:
    parameters: [{ $ref: "#/components/parameters/id" }]
    put:
      tags: [notifications]
      summary: Kanalı güncelle (editör+)
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/NotificationInput" } } } }
      responses: { "200": { description: Kanal } }
    delete: { tags: [notifications], summary: Kanalı sil (editör+), responses: { "200": { $ref: "#/components/responses/OK" } } }
  /api/notifications/test:
    post:
      tags: [notifications]
      summary: Formdaki ayarla test bildirimi (editör+; dakikada 10)
      requestBody: { content: { application/json: { schema: { type: object, properties: { id: { type: integer, format: int64 }, type: { type: string }, config: { type: object, additionalProperties: true }, lang: { type: string } } } } } }
      responses: { "200": { $ref: "#/components/responses/OK" }, "422": { $ref: "#/components/responses/Error" } }
  /api/notifications/{id}/samples:
    post: { tags: [notifications], summary: Her türden örnek bildirim gönder (editör+; dakikada bir), parameters: [{ $ref: "#/components/parameters/id" }], responses: { "202": { description: "{count}" } } }
...
  /api/push/{token}:
    get:
      tags: [public]
      summary: Push monitörü sinyali (cron / betik)
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [up, down] } }
        - { name: msg, in: query, schema: { type: string, maxLength: 250 } }
        - { name: ping, in: query, schema: { type: integer } }
      responses: { "200": { $ref: "#/components/responses/OK" }, "404": { $ref: "#/components/responses/Error" }, "409": { description: Monitör durdurulmuş }, "429": { $ref: "#/components/responses/Error" } }
    post:
      tags: [public]
      summary: Push sinyali (POST)
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses: { "200": { $ref: "#/components/responses/OK" } }
  /api/servers:
    get: { tags: [servers], summary: Sunucu ajanları (son metriklerle), responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Server" } } } } } } }
    post: { tags: [servers], summary: Sunucu ekle (yönetici); kurulum komutu ve token döner, requestBody: { content: { application/json: { schema: { type: object, properties: { name: { type: string } } } } } }, responses: { "201": { description: Sunucu + token } } }
  /api/servers/{id}:
    get: { tags: [servers], summary: Sunucu ayrıntısı (host bilgisi, son örnek, uyarılar, konteynerler), parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { description: Sunucu, content: { application/json: { schema: { $ref: "#/components/schemas/Server" } } } } } }
...
  /api/servers/{id}/events:
    get: { tags: [servers], summary: Uyarı olayları, parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { description: "{events[]}" } } }
  /api/servers/{id}/incidents:
    get: { tags: [servers], summary: Sunucunun olayları, parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Incident" } } } } } } }
...
  /api/servers/{id}/notifications:
    put:
      tags: [servers]
      summary: Sunucunun bildirim kanalları (yönetici)
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { notification_ids: { type: array, items: { type: integer, format: int64 } }, notification_bindings: { type: array, items: { type: object, additionalProperties: true } } } } } } }
      responses: { "200": { description: Bağlantılar } }
...
  /api/users:
    get: { tags: [users], summary: Kullanıcılar (yönetici), responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/UserRecord" } } } } } } }
    post: { tags: [users], summary: Kullanıcı ekle (yönetici), requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/UserInput" } } } }, responses: { "201": { description: Kullanıcı }, "409": { $ref: "#/components/responses/Error" } } }
  /api/users/{id}:
    parameters: [{ $ref: "#/components/parameters/id" }]
    put: { tags: [users], summary: Kullanıcıyı güncelle (yönetici), requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/UserInput" } } } }, responses: { "200": { description: Kullanıcı } } }
    delete: { tags: [users], summary: Kullanıcıyı sil (yönetici), responses: { "200": { $ref: "#/components/responses/OK" } } }
  /api/users/{id}/password:
    post: { tags: [users], summary: Geçici şifre ver (oturumları kapatır; yönetici), parameters: [{ $ref: "#/components/parameters/id" }], requestBody: { content: { application/json: { schema: { type: object, properties: { password: { type: string } } } } } }, responses: { "200": { $ref: "#/components/responses/OK" } } }
...
  /api/settings:
    get: { tags: [settings], summary: Uygulama ayarları (yönetici), responses: { "200": { description: Ayarlar, content: { application/json: { schema: { $ref: "#/components/schemas/Settings" } } } } } }
    put: { tags: [settings], summary: Ayarları kaydet (GET yanıtı aynen gönderilebilir), requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/Settings" } } } }, responses: { "200": { description: Ayarlar }, "400": { $ref: "#/components/responses/Error" } } }
...
  /api/audit:
    get:
      tags: [settings]
      summary: İşlem kaydı (yönetici)
      parameters:
        - { name: user, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string }, description: "Tam kod (monitor.delete) ya da alan öneki (monitor.)" }
        - { name: from, in: query, schema: { type: integer } }
        - { name: to, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
        - { name: before, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, maximum: 500 } }
      responses: { "200": { description: Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/AuditEntry" } } } } } }
...
  /api/events:
    get:
      tags: [monitors]
      summary: Canlı olay akışı (Server-Sent Events)
      description: "`text/event-stream`: beat, incident, maintenance, server olayları; kısıtlı kullanıcıda yalnızca izinli monitörler."
      parameters: [{ name: lang, in: query, schema: { $ref: "#/components/schemas/Lang" } }]
      responses: { "200": { description: Akış, content: { text/event-stream: {} } }, "429": { $ref: "#/components/responses/Error" } }
...